recorder: improve when "file-{created,completed}" signals are emitted
Emit the signal when "splitmuxsink-fragment-{opened,closed}" message
appears on GstBus so that we don't have to rely on "format-location"
signal in order to detect the change of output file.

Simplify the code by removing hwangsae_recorder_handle_new_file() and
leave it to the splitmuxsink itself to generate filename for each
fragment of the recording.