relay: add "authenticate" signal
HwangsaeRelay will emit "authenticate" when a new client is about to
connect. Its callback is passed the username and resource from the
connection's Stream ID and the peer's address.

The application should return TRUE from "authenticate" callback in order
to accept the incoming connection or FALSE to reject it.

There may be multiple listeners connected to this signal. The connection
gets rejectes if any of them returns FALSE.

When nothing is connected to "authenticate", the relay will accept all
connections.